requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.35.0
decorator==4.4.2
deepgram-sdk==2.11.0
ffmpeg==1.4
Flask==3.0.0
frozenlist==1.4.0
idna==3.4
imageio==2.31.4
imageio-ffmpeg==0.4.9
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
moviepy==1.0.3
multidict==6.0.4
numpy==1.26.0
Pillow==10.0.1
proglog==0.1.10
python-dotenv==1.0.0
requests==2.31.0
six==1.16.0
tqdm==4.66.1
urllib3==1.26.16
websockets==11.0.3
Werkzeug==3.0.0
yarl==1.9.2
